resolve comments + add prefix
