+++ solutions/fully-configurable/README.md
@@ -56,6 +56,7 @@ No resources.
 | <a name="input_other_sans"></a> [other\_sans](#input\_other\_sans) | The custom Object Identifier (OID) or UTF8-string Subject Alternative Names (SANs) to define for the CA certificate. The alternative names must match the values that are specified in the 'allowed\_other\_sans' field in the associated certificate template. [Learn More](https://cloud.ibm.com/docs/secrets-manager?topic=secrets-manager-secrets-manager-cli#secrets-manager-configurations-cli) | `list(string)` | `null` | no |
 | <a name="input_permitted_dns_domains"></a> [permitted\_dns\_domains](#input\_permitted\_dns\_domains) | The Allowed DNS domains or subdomains for the certificates which is to be signed and issued by the CA certificate. | `list(string)` | `null` | no |
 | <a name="input_postal_code"></a> [postal\_code](#input\_postal\_code) | The Street Address values that are defined in the subject field of the resulting certificate. [Learn More](https://cloud.ibm.com/docs/secrets-manager?topic=secrets-manager-secrets-manager-cli#secrets-manager-configurations-cli) | `list(string)` | `null` | no |
+| <a name="input_prefix"></a> [prefix](#input\_prefix) | The prefix to add to all resources created by this solution. To not use any prefix value, you can set this value to `null` or an empty string. | `string` | n/a | yes |
 | <a name="input_private_key_format"></a> [private\_key\_format](#input\_private\_key\_format) | The format in which the private key is generated. | `string` | `"der"` | no |
 | <a name="input_province"></a> [province](#input\_province) | The Province (ST) values that are defined in the subject field of the resulting certificate. [Learn More](https://cloud.ibm.com/docs/secrets-manager?topic=secrets-manager-secrets-manager-cli#secrets-manager-configurations-cli) | `list(string)` | `null` | no |
 | <a name="input_return_format"></a> [return\_format](#input\_return\_format) | The format in which the data is returned. | `string` | `"pem"` | no |
