+++ solutions/fully-configurable/catalogValidationValues.json.template
@@ -1,4 +1,5 @@
 {
   "ibmcloud_api_key": $VALIDATION_APIKEY,
-  "existing_secrets_manager_crn": $SM_CRN
+  "existing_secrets_manager_crn": $SM_CRN,
+  "prefix": $PREFIX
 }
